+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 # CHANGELOG
 
-## 0.16.3dev
+## 0.16.5dev
+* Style improvements to DAG plot
+* DAG plot only includes task names by default (#393)
+* `ploomber plot --include-products/-p` generates plots with task names and products
+* `DAG.plot(include_products=True)` generates plots with task names and products
+
+## 0.16.4 (2022-03-11)
+* Using UTF-8 for reading and writing in notebook tasks ([#334](https://github.com/ploomber/ploomber/issues/334))
+
+## 0.16.3 (2022-03-06)
+* Clearer error message when DAG deepcopy fails
+* Beta release of cloud pipeline monitoring
+* More robust suggestions when invoking a non-existing command
+* CLI loading performance improvements
+* Prints message before starting to load the pipeline for better user feedback
+* Displaying community link when DAG fails to render or build
 
 ## 0.16.2 (2022-03-03)
 * Improved documentation in "ploomber nb --help" ([#623](https://github.com/ploomber/ploomber/issues/623))